+++ pdf/pdf_image.c
@@ -2019,7 +2019,7 @@ pdfi_do_image(pdf_context *ctx, pdf_dict *page_dict, pdf_dict *stream_dict, pdf_
      */
     if (mask_stream) {
         /* Calculate expected mask size */
-        mask_size = ((((t3image.MaskDict.BitsPerComponent * t3image.MaskDict.Width) + 7) / 8) * t3image.MaskDict.Height);
+        mask_size = ((((t3image.MaskDict.BitsPerComponent * (int64_t)t3image.MaskDict.Width) + 7) / 8) * (int64_t)t3image.MaskDict.Height);
         code = pdfi_stream_to_buffer(ctx, mask_stream, &mask_buffer, (int64_t *)&mask_size);
         if (code < 0)
             goto cleanupExit;
